# openapi/generated/platform-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enterprise RAG Platform API",
    "description": "知识空间、文档接入和平台治理 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/uploads": {
      "post": {
        "operationId": "createUploadSession",
        "summary": "创建浏览器直传会话",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUploadSessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "上传会话和短时预签名 URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadSessionEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "身份缺失或无法验证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行操作",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "文件数量或大小超过配置上限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/uploads/{uploadId}": {
      "get": {
        "operationId": "getUploadSession",
        "summary": "恢复上传会话",
        "parameters": [
          {
            "name": "uploadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "上传会话",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadSessionEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "身份缺失或无法验证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行操作",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "cancelUploadSession",
        "summary": "取消上传会话",
        "parameters": [
          {
            "name": "uploadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "已取消"
          },
          "401": {
            "description": "身份缺失或无法验证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行操作",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/{uploadId}/parts": {
      "post": {
        "operationId": "createUploadParts",
        "summary": "按需签发 Multipart 分片 URL",
        "parameters": [
          {
            "name": "uploadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUploadPartsRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "分片上传指令",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadPartListEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "身份缺失或无法验证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行操作",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/{uploadId}/complete": {
      "post": {
        "operationId": "completeUpload",
        "summary": "HEAD 验证对象并原子创建文档与任务事实",
        "parameters": [
          {
            "name": "uploadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "文档、版本、文件和任务",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteUploadEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "身份缺失或无法验证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行操作",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "对象事实不匹配",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs": {
      "get": {
        "operationId": "listIngestionJobs",
        "summary": "按权限游标分页列出入库任务",
        "responses": {
          "200": {
            "description": "任务列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionJobListEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "身份缺失或无法验证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行操作",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{jobId}": {
      "get": {
        "operationId": "getIngestionJob",
        "summary": "读取任务与步骤进度",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 300
            }
          }
        ],
        "responses": {
          "200": {
            "description": "任务详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionJobEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "身份缺失或无法验证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行操作",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{jobId}/cancel": {
      "post": {
        "operationId": "cancelIngestionJob",
        "summary": "取消非终态任务",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 300
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelIngestionJobRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "已取消任务",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionJobEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "身份缺失或无法验证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行操作",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{jobId}/events": {
      "get": {
        "operationId": "streamIngestionJobEvents",
        "summary": "使用 SSE 和 Last-Event-ID 续传任务事件",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 300
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "text/event-stream 任务事件"
          },
          "401": {
            "description": "身份缺失或无法验证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行操作",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{jobId}/events/poll": {
      "get": {
        "operationId": "pollIngestionJobEvents",
        "summary": "使用 ETag 和游标轮询任务事件",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 300
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "任务事件",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionJobEventListEnvelope"
                }
              }
            }
          },
          "304": {
            "description": "没有新事件"
          },
          "401": {
            "description": "身份缺失或无法验证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "当前身份无权执行操作",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "trustedUserHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Authenticated-User"
      },
      "mockPreset": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RAG-Mock-User"
      }
    },
    "schemas": {
      "ApiError": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "minLength": 1
          },
          "traceId": {
            "type": "string",
            "minLength": 1
          },
          "code": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "VALIDATION_ERROR",
                  "NOT_FOUND",
                  "SERVICE_UNAVAILABLE",
                  "DEADLINE_EXCEEDED",
                  "INTERNAL_ERROR"
                ]
              },
              {
                "type": "string",
                "minLength": 1
              }
            ]
          },
          "message": {
            "type": "string",
            "minLength": 1
          },
          "retryable": {
            "type": "boolean"
          },
          "details": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          }
        },
        "required": [
          "requestId",
          "code",
          "message",
          "retryable"
        ],
        "additionalProperties": false
      },
...
      "CreateUploadSessionRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "spaceId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "files": {
            "minItems": 1,
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "clientFileId": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 100
                },
                "originalFileName": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 512
                },
                "sizeBytes": {
                  "type": "integer",
                  "exclusiveMinimum": 0,
                  "maximum": 9007199254740991
                },
                "contentType": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 160
                },
                "sha256": {
                  "type": "string",
                  "pattern": "^[a-f0-9]{64}$"
                }
              },
              "required": [
                "clientFileId",
                "originalFileName",
                "sizeBytes",
                "contentType"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "spaceId",
          "files"
        ],
        "additionalProperties": false
      },
...
      "CreateUploadPartsRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "fileId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "partNumbers": {
            "minItems": 1,
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10000
            }
          }
        },
        "required": [
          "fileId",
          "partNumbers"
        ],
        "additionalProperties": false
      },
...
      "UploadSessionEnvelope": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "minLength": 1
          },
          "traceId": {
            "type": "string",
            "minLength": 1
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "spaceId": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "status": {
                "type": "string",
                "enum": [
                  "ACTIVE",
                  "COMPLETED",
                  "CANCELLED",
                  "EXPIRED"
                ]
              },
              "expiresAt": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
              },
              "files": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "fileId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "clientFileId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "originalFileName": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 240
                    },
                    "sizeBytes": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    },
                    "contentType": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 160
                    },
                    "strategy": {
                      "type": "string",
                      "enum": [
                        "SINGLE",
                        "MULTIPART"
                      ]
                    },
                    "partSizeBytes": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    },
                    "partCount": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    },
                    "uploadUrl": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "uri"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    },
                    "completed": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "fileId",
                    "clientFileId",
                    "originalFileName",
                    "sizeBytes",
                    "contentType",
                    "strategy",
                    "partSizeBytes",
                    "partCount",
                    "uploadUrl",
                    "expiresAt",
                    "completed"
                  ],
                  "additionalProperties": false
                }
              }
            },
            "required": [
              "id",
              "spaceId",
              "status",
              "expiresAt",
              "createdAt",
              "files"
            ],
            "additionalProperties": false
          },
          "meta": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          }
        },
        "required": [
          "requestId",
          "data"
        ],
        "additionalProperties": false
      },
      "UploadPartListEnvelope": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "minLength": 1
          },
          "traceId": {
            "type": "string",
            "minLength": 1
          },
          "data": {
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "partNumber": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    },
                    "uploadUrl": {
                      "type": "string",
                      "format": "uri"
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    }
                  },
                  "required": [
                    "partNumber",
                    "uploadUrl",
                    "expiresAt"
                  ],
                  "additionalProperties": false
                }
              }
            },
            "required": [
              "items"
            ],
            "additionalProperties": false
          },
          "meta": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          }
        },
        "required": [
          "requestId",
          "data"
        ],
        "additionalProperties": false
      },
...
    }
  }
}